requirements.txt
fastapi==0.86.0
uvicorn==0.19.0
starlette~=0.20.4
numpy~=1.23.4
Pillow~=9.3.0
locust~=2.13.0
python-multipart==0.0.5